Disassembler: Decode SIMD ldr/str instructions, bugref:10394
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@106746 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ include/VBox/dis-armv8.h
@@ -111,7 +111,7 @@ typedef struct
     union
     {
         /** Offset from the base register. */
-        int16_t                     offBase;
+        int32_t                     offBase;
         /** Amount of bits to extend. */
         uint8_t                     cExtend;
     } u;
